Create a URL board with a manifest

Create a URL board with a manifest

Usage

board_url_manifest(url, cache = NULL, use_cache_on_failure = is_interactive())

Arguments

url

root URL of the board

cache

Cache path. Every board requires a local cache to avoid downloading files multiple times. The default stores in a standard cache location for your operating system, but you can override if needed.

use_cache_on_failure

If the pin fails to download, is it ok to use the last cached version? Defaults to is_interactive() so you'll be robust to poor internet connectivity when exploring interactively, but you'll get clear errors when the code is deployed.

Examples

board <- board_url_manifest(url_demo_manifest())
pins::pin_list(board)
#> [1] "mtcars-csv"  "mtcars-json"
pins::pin_meta(board, "mtcars-csv")
#> List of 11
#>  $ file       : chr "mtcars-csv.csv"
#>  $ file_size  : 'fs_bytes' int 1.27K
#>  $ pin_hash   : chr "48c73eef04b242dd"
#>  $ type       : chr "csv"
#>  $ title      : chr "mtcars-csv: a pinned 32 x 11 data frame"
#>  $ description: NULL
#>  $ created    : POSIXct[1:1], format: "2022-08-11 15:58:05"
#>  $ api_version: num 1
#>  $ user       : list()
#>  $ name       : chr "mtcars-csv"
#>  $ local      :List of 4
#>   ..$ dir     : 'fs_path' chr "~/.cache/pins/url/4634077d3a3c2fa97df429def01d8a31"
#>   ..$ url     : chr "https://raw.githubusercontent.com/ijlyttle/pinsManifest/main/tests/testthat/pins/mtcars-csv/20220811T155805Z-48c73/"
#>   ..$ version : NULL
#>   ..$ file_url: chr "https://raw.githubusercontent.com/ijlyttle/pinsManifest/main/tests/testthat/pins/mtcars-csv/20220811T155805Z-48"| __truncated__